Got multiple files ecosystem up and running
