+++ keyboards/lily58l/keymaps/georg/oled.c
@@ -1,8 +1,5 @@
-#ifndef OLED_C
-#define OLED_C
-
 #include QMK_KEYBOARD_H
-#include "keymap.c"
+#include "keymap.h"
 
 #ifdef OLED_DRIVER_ENABLE
 oled_rotation_t oled_init_user(oled_rotation_t rotation) {
@@ -67,6 +64,4 @@ void oled_task_user(void) {
         render_right(); 
     }
 }
-#endif
-
 #endif
