+++ keyboards/lily58l/keymaps/georg/rules.mk
@@ -2,7 +2,7 @@ OLED_DRIVER_ENABLE = yes   # Enables the use of OLED displays
 ENCODER_ENABLE = yes       # ENables the use of one or more encoders
 RGBLIGHT_ENABLE = yes      # Enable keyboard RGB light
 WPM_ENABLE = yes	
-SRC += oled.c
+SRC += oled.c encoder.c
 
 # uncomment to the line to customize the setting
 #define ENCODER_DIRECTION_FLIP # Switch encoder’s directions
